extracting functions WIP
